dos_concise/requirements.txt
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.21
langsmith==0.0.87
openai==1.9.0
faiss-cpu==1.7.4
pypdf==4.0.1
tiktoken==0.5.2
python-dotenv==1.0.1
httpx==0.27.2
